+++ src/re_com/tabs.cljs
@@ -31,6 +31,7 @@
 (def horizontal-tabs-args-desc
   (when include-args-desc?
     [{:name :model            :required true                  :type "unique-id | r/atom"                                                  :description "the unique identifier of the currently selected tab"}
+     {:name :disabled?        :required false                 :type "boolean | r/atom"                                                    :description "disables all tabs."}
      {:name :tabs             :required true                  :type "vector of tabs | r/atom" :validate-fn vector-of-maps?                :description "one element in the vector for each tab. Typically, each element is a map with :id and :label keys"}
      {:name :on-change        :required true                  :type "unique-id -> nil"        :validate-fn fn?                            :description "called when user alters the selection. Passed the unique identifier of the selection"}
      {:name :id-fn            :required false :default :id    :type "tab -> anything"         :validate-fn ifn?                           :description [:span "given an element of " [:code ":tabs"] ", returns its unique identifier (aka id)"]}
@@ -43,13 +44,14 @@
      {:name :debug-as         :required false                 :type "map"                     :validate-fn map?                           :description [:span "Used in dev builds to assist with debugging, when one component is used implement another component, and we want the implementation component to masquerade as the original component in debug output, such as component stacks. A map optionally containing keys" [:code ":component"] "and" [:code ":args"] "."]}]))
 
 (defn horizontal-tabs
-  [& {:keys [model tabs on-change id-fn label-fn class style attr parts]
+  [& {:keys [model tabs on-change id-fn label-fn class style attr parts disabled?]
       :or   {id-fn :id label-fn :label}
       :as   args}]
   (or
    (validate-args-macro horizontal-tabs-args-desc args)
    (let [current  (deref-or-value model)
          tabs     (deref-or-value tabs)
+         disabled? (deref-or-value disabled?)
          _        (assert (not-empty (filter #(= current (id-fn %)) tabs)) "model not found in tabs vector")]
      [:ul
       (merge {:class (str "nav nav-tabs noselect rc-tabs " class)
@@ -60,10 +62,13 @@
       (for [t tabs]
         (let [id        (id-fn  t)
               label     (label-fn  t)
+              disabled? (or disabled? (:disabled? t))
               selected? (= id current)]                   ;; must use current instead of @model to avoid reagent warnings
           [:li
            (merge
-            {:class (str (if selected? "active rc-tab ") (get-in parts [:tab :class]))
+            {:class (str (when disabled? "disabled ")
+                         (when selected? "active rc-tab ")
+                         (get-in parts [:tab :class]))
              :style (get-in parts [:tab :style])
              :key   (str id)}
             (get-in parts [:tab :attr]))
@@ -72,7 +77,7 @@
              {:class    (str "rc-tab-anchor " (get-in parts [:anchor :class]))
               :style    (merge {:cursor "pointer"}
                                style)
-              :on-click (when (and on-change (not selected?)) (handler-fn (on-change id)))}
+              :on-click (when (and on-change (not selected?) (not disabled?)) (handler-fn (on-change id)))}
              (get-in parts [:anchor :attr]))
             label]]))])))
 
@@ -104,15 +109,16 @@
       {:name :parts            :required false                        :type "map"                    :validate-fn (parts? bar-tabs-parts) :description "See Parts section below."}))))
 
 (defn- bar-tabs
-  [& {:keys [model tabs on-change id-fn label-fn tooltip-fn tooltip-position vertical? class style attr parts validate?] :as args}]
+  [& {:keys [model tabs on-change id-fn label-fn tooltip-fn tooltip-position vertical? class style attr parts validate? disabled?] :as args}]
   (let [showing (reagent/atom nil)]
-    (fn [& {:keys [model tabs]}]
+    (fn [& {:keys [model tabs disabled?]}]
       (let [current  (deref-or-value model)
             tabs     (deref-or-value tabs)
+            disabled? (deref-or-value disabled?)
             _        (assert (or (not validate?) (not-empty (filter #(= current (id-fn %)) tabs))) "model not found in tabs vector")]
         (into [:div
                (merge
-                {:class (str "noselect btn-group" (if vertical? "-vertical") " rc-tabs " class)
+                {:class (str "noselect btn-group" (when vertical? "-vertical") " rc-tabs " class)
                  :style (merge (flex-child-style "none")
                                (get-in parts [:wrapper :style]))}
                 (->attr args)
@@ -122,19 +128,25 @@
                       label     (label-fn t)
                       tooltip   (when tooltip-fn (tooltip-fn t))
                       selected? (= id current)
-                      the-button [:button
-                                  (merge
-                                   {:type     "button"
-                                    :key      (str id)
-                                    :class    (str "btn btn-default " (if selected? "active ") "rc-tabs-btn " (get-in parts [:button :class]))
-                                    :style    style
-                                    :on-click (when (and on-change (not selected?))
-                                                (handler-fn (on-change id)))}
-                                   (when tooltip
-                                     {:on-mouse-over (handler-fn (reset! showing id))
-                                      :on-mouse-out  (handler-fn (swap! showing #(when-not (= id %) %)))})
-                                   (get-in parts [:button :attr]))
-                                  label]]
+                      disabled? (or disabled? (:disabled? t))
+                      the-button
+                      [:button
+                       (merge
+                        {:type     "button"
+                         :key      (str id)
+                         :class    (str "btn btn-default "
+                                        (when disabled? "disabled ")
+                                        (when selected? "active ")
+                                        "rc-tabs-btn "
+                                        (get-in parts [:button :class]))
+                         :style    style
+                         :on-click (when (and on-change (not selected?) (not disabled?))
+                                     (handler-fn (println (gensym)) (on-change id)))}
+                        (when tooltip
+                          {:on-mouse-over (handler-fn (reset! showing id))
+                           :on-mouse-out  (handler-fn (swap! showing #(when-not (= id %) %)))})
+                        (get-in parts [:button :attr]))
+                       label]]
                   (if tooltip
                     [popover-tooltip
                      :src      (at)
@@ -148,7 +160,7 @@
                     the-button))))))))
 
 (defn horizontal-bar-tabs
-  [& {:keys [model tabs on-change id-fn label-fn tooltip-fn tooltip-position class style attr parts src debug-as validate?]
+  [& {:keys [model tabs on-change id-fn label-fn tooltip-fn tooltip-position class style attr parts src debug-as validate? disabled?]
       :or   {id-fn :id label-fn :label tooltip-fn :tooltip}
       :as   args}]
   (or
@@ -156,6 +168,7 @@
    (bar-tabs
     :model            model
     :tabs             tabs
+    :disabled?        disabled?
     :on-change        on-change
     :id-fn            id-fn
     :label-fn         label-fn
@@ -171,7 +184,7 @@
     :validate?        validate?)))
 
 (defn vertical-bar-tabs
-  [& {:keys [model tabs on-change id-fn label-fn class style attr parts src debug-as validate?]
+  [& {:keys [model tabs on-change id-fn label-fn class style attr parts src debug-as validate? disabled?]
       :or   {id-fn :id label-fn :label}
       :as   args}]
   (or
@@ -179,6 +192,7 @@
    (bar-tabs
     :model     model
     :tabs      tabs
+    :disabled? disabled?
     :on-change on-change
     :id-fn     id-fn
     :label-fn  label-fn
@@ -216,10 +230,11 @@
       {:name :parts            :required false                        :type "map"                      :validate-fn (parts? pill-tabs-parts) :description "See Parts section below."}))))
 
 (defn- pill-tabs    ;; tabs-like in action
-  [& {:keys [model tabs on-change id-fn label-fn vertical? class style attr parts src] :as args}]
-  (let [current  (deref-or-value model)
-        tabs     (deref-or-value tabs)
-        _        (assert (not-empty (filter #(= current (id-fn %)) tabs)) "model not found in tabs vector")]
+  [& {:keys [model tabs on-change id-fn label-fn vertical? class style attr parts src disabled?] :as args}]
+  (let [current   (deref-or-value model)
+        tabs      (deref-or-value tabs)
+        disabled? (deref-or-value disabled?)
+        _         (assert (not-empty (filter #(= current (id-fn %)) tabs)) "model not found in tabs vector")]
     [:ul
      (merge
       {:class (str "rc-tabs noselect nav nav-pills" (when vertical? " nav-stacked") " " class)
@@ -231,24 +246,29 @@
      (for [t tabs]
        (let [id        (id-fn  t)
              label     (label-fn  t)
+             disabled? (or disabled? (:disabled? t))
              selected? (= id current)]                   ;; must use 'current' instead of @model to avoid reagent warnings
          [:li
           (merge
-           {:class    (str "rc-tabs-pill " (if selected? "active " "") (get-in parts [:tab :class]))
-            :style    (get-in parts [:tab :style])
-            :key      (str id)}
+           {:class (str "rc-tabs-pill "
+                        (when disabled? "disabled ")
+                        (when selected? "active ")
+                        (get-in parts [:tab :class]))
+            :style (get-in parts [:tab :style])
+            :key   (str id)}
            (get-in parts [:tab :attr]))
           [:a
            (merge
-            {:class     (str "rc-tabs-anchor " (get-in parts [:anchor :class]))
-             :style     (merge {:cursor "pointer"}
-                               style)
-             :on-click  (when (and on-change (not selected?)) (handler-fn (on-change id)))}
+            {:class    (str "rc-tabs-anchor " (get-in parts [:anchor :class]))
+             :style    (merge {:cursor "pointer"}
+                              style)
+             :on-click (when (and on-change (not selected?) (not disabled?))
+                         (handler-fn (on-change id)))}
             (get-in parts [:anchor :attr]))
            label]]))]))
 
 (defn horizontal-pill-tabs
-  [& {:keys [model tabs on-change id-fn label-fn class style attr parts src debug-as]
+  [& {:keys [model tabs on-change id-fn label-fn class style attr parts src debug-as disabled?]
       :or   {id-fn :id label-fn :label}
       :as   args}]
   (or
@@ -256,6 +276,7 @@
    (pill-tabs
     :model     model
     :tabs      tabs
+    :disabled? disabled?
     :on-change on-change
     :id-fn     id-fn
     :label-fn  label-fn
@@ -268,7 +289,7 @@
     :debug-as  debug-as)))
 
 (defn vertical-pill-tabs
-  [& {:keys [model tabs on-change id-fn label-fn class style attr parts src debug-as]
+  [& {:keys [model tabs on-change id-fn label-fn class style attr parts src debug-as disabled?]
       :or   {id-fn :id label-fn :label}
       :as   args}]
   (or
@@ -276,6 +297,7 @@
    (pill-tabs
     :model     model
     :tabs      tabs
+    :disabled? disabled?
     :on-change on-change
     :id-fn     id-fn
     :label-fn  label-fn